requirements.txt
pika>=1.3.0
pycryptodome>=3.20.0
